renamed pipInstall to install
