+++ packages/now-python/src/install.ts
